build: Configure Jacoco
Resolves: DCMAW-7374
