+++ buildLogic/plugins/src/main/kotlin/sonarqube-root-config.gradle.kts
